+++ Editor/guis/LoudnessCorrectionFilterGUI.h
@@ -65,6 +65,7 @@ private slots:
 	void on_manualVolumeCheckBox_toggled(bool checked);
 	void on_volumeSpinBox_valueChanged(double value);
 	void on_fastEngineCheckBox_toggled(bool checked);
+	void on_takeoverVolumeCheckBox_toggled(bool checked);
 	void on_volumeFollowComboBox_currentIndexChanged(int index);
 	void on_studioButton_clicked();
 	void on_calibrateButton_clicked();